+++ src/sio/mod.rs
@@ -1,2 +1,3 @@
 pub mod client;
 pub mod metrics;
+pub mod utils;